[#32] Refactored server file
